Moved files into Editor and Story namespaces
