+++ S2VX.Game/Story/Note.cs
@@ -2,7 +2,7 @@
 using osu.Framework.Utils;
 using osuTK;
 
-namespace S2VX.Game
+namespace S2VX.Game.Story
 {
     public class Note : RelativeBox
     {
@@ -10,7 +10,7 @@ namespace S2VX.Game
         public Vector2 Coordinates { get; set; } = Vector2.Zero;
 
         [Resolved]
-        private Story story { get; set; } = null;
+        private S2VXStory story { get; set; } = null;
 
         [BackgroundDependencyLoader]
         private void load()
